Adds Exceptions, models, and minor fixes.
This commit adds custom exceptions for various errors that can be
thrown in the L5SimpleFMBase class. It also adds a BaseModel class and
an example of how to extend the class to create your own model.

It also includes a couple of minor fixes for the base class and the
addition of several other command wrappers (max, skip, etc)