Adds the base model method resetLayout
This gives the developer the option of changing the model’s layout at
runtime. This can be an advantage because when you make a request from
FileMaker via CWP, you have no control over what fields are returned,
i.e. there is no equivalent to `Select id,name,…` from FIleMaker every
request is a `Select *…`.

With the addition of resetLayout the developer can use multiple layouts
to represent one entity and switch to a layout that is not the default
(the one defined in the model) at runtime.